+++ custom_components/houseplan/const.py
@@ -45,7 +45,7 @@ PLAN_ORPHAN_TTL_S = 3600
 SCHEDULED_GRACE_S = 30 * 24 * 3600
 FILES_DIR = "houseplan/files"
 CONF_ADMIN_ONLY = "admin_only"
-VERSION = "1.61.0-beta.2"
+VERSION = "1.61.0-beta.3"
 
 DEFAULT_CONFIG: dict = {
     "spaces": [],